+++ lib/authly-domain/src/lib.rs
@@ -1,3 +1,5 @@
+use serde::{Deserialize, Serialize};
+
 pub mod access_control;
 pub mod access_token;
 pub mod audit;
@@ -12,7 +14,11 @@ pub mod extract;
 pub mod id;
 pub mod instance;
 pub mod login;
+pub mod migration;
 pub mod persona_directory;
 pub mod repo;
 pub mod session;
 pub mod tls;
+
+#[derive(Clone, Copy, Serialize, Deserialize, Debug)]
+pub struct IsLeaderDb(pub bool);
