+++ lib/authly-domain/src/migration.rs
@@ -1,0 +1,3 @@
+#[derive(rust_embed::Embed)]
+#[folder = "migrations"]
+pub struct Migrations;
