+++ lib/authly-domain/src/repo/mod.rs
@@ -1,4 +1,6 @@
 pub mod crypto_repo;
+pub mod directory_repo;
 pub mod entity_repo;
+pub mod init_repo;
 pub mod object_repo;
 pub mod session_repo;
