+++ src/db/directory_db.rs
@@ -2,7 +2,7 @@
 
 use std::collections::HashMap;
 
-use authly_common::id::{AnyId, AttrId, DirectoryId, PolicyId, PropId, ServiceId};
+use authly_common::id::{AnyId, AttrId, DirectoryId, PolicyId, PropId};
 use authly_db::{param::ToBlob, params, Db, DbResult, FromRow, Row, TryFromRow};
 use authly_domain::directory::DirKey;
 use indoc::indoc;
@@ -14,14 +14,6 @@ use super::{
     policy_db::DbPolicy,
     service_db::{NamespaceProperty, PropertyKind},
 };
-
-pub async fn query_dir_key(deps: &impl Db, dir_id: DirectoryId) -> DbResult<Option<DirKey>> {
-    deps.query_map_opt::<DirKey>(
-        "SELECT key FROM directory WHERE id = $1".into(),
-        params!(dir_id.to_blob()),
-    )
-    .await
-}
 
 pub struct DbDirectory {
     pub key: DirKey,
@@ -77,29 +69,6 @@ impl DbDirectoryNamespaceLabel {
         deps.query_map(
             // FIXME: unindexed query
             "SELECT id, label FROM namespace WHERE dir_key = $1".into(),
-            params!(dir_key.0),
-        )
-        .await
-    }
-}
-
-pub struct DbDirectoryService {
-    pub svc_eid: ServiceId,
-}
-
-impl FromRow for DbDirectoryService {
-    fn from_row(row: &mut impl Row) -> Self {
-        Self {
-            svc_eid: row.get_id("svc_eid"),
-        }
-    }
-}
-
-impl DbDirectoryService {
-    pub async fn query(deps: &impl Db, dir_key: DirKey) -> DbResult<Vec<Self>> {
-        deps.query_map(
-            // FIXME: unindexed query
-            "SELECT svc_eid FROM svc WHERE dir_key = $1".into(),
             params!(dir_key.0),
         )
         .await
