+++ src/document/doc_compiler.rs
@@ -12,13 +12,14 @@ use authly_common::{
 };
 use authly_db::{Db, DbError};
 use authly_domain::ctx::KubernetesConfig;
+use authly_domain::repo::directory_repo::query_dir_key;
 use authly_domain::{ctx::GetDb, directory::DirKey, id::BuiltinProp};
 use serde::de::value::StrDeserializer;
 use serde::Deserialize;
 use serde_spanned::Spanned;
 use tracing::debug;
 
-use crate::db::directory_db::{query_dir_key, DbDirectoryNamespaceLabel, DbDirectoryPolicy};
+use crate::db::directory_db::{DbDirectoryNamespaceLabel, DbDirectoryPolicy};
 use crate::db::policy_db::DbPolicy;
 use crate::db::{directory_db, policy_db, service_db, Identified};
 use crate::document::compiled_document::{
