+++ src/encryption.rs
@@ -1,73 +1,30 @@
-use std::{collections::HashMap, time::Duration};
+use std::time::Duration;
 
-use aes_gcm_siv::{
-    aead::{Aead, Nonce},
-    Aes256GcmSiv, KeyInit,
-};
+use aes_gcm_siv::{aead::Nonce, Aes256GcmSiv};
 use anyhow::anyhow;
-use authly_common::id::PropId;
 use authly_db::Db;
 use authly_domain::{
-    encryption::{random_nonce, AesKey, DecryptedDeks},
+    encryption::{gen_prop_deks, AesKey, DecryptedDeks, DecryptedMaster, MasterVersion},
     id::BuiltinProp,
+    repo::crypto_repo,
+    IsLeaderDb,
 };
 use authly_secrets::AuthlySecrets;
-use rand::{rngs::OsRng, RngCore};
 use secrecy::ExposeSecret;
-use time::OffsetDateTime;
 use tracing::info;
-use zeroize::{Zeroize, Zeroizing};
-
-use crate::{db::cryptography_db, IsLeaderDb};
-
-#[derive(Clone)]
-pub struct MasterVersion {
-    pub version: Vec<u8>,
-    pub created_at: time::OffsetDateTime,
-}
-
-pub struct DecryptedMaster {
-    encrypted: MasterVersion,
-    key: AesKey,
-}
-
-impl DecryptedMaster {
-    pub fn fake_for_test() -> Self {
-        Self {
-            encrypted: MasterVersion {
-                version: vec![],
-                created_at: OffsetDateTime::now_utc(),
-            },
-            key: {
-                let mut key = [0u8; 32];
-                OsRng.fill_bytes(key.as_mut_slice());
-
-                AesKey::new(key.into())
-            },
-        }
-    }
-}
-
-#[derive(Clone)]
-pub struct EncryptedDek {
-    pub nonce: Nonce<Aes256GcmSiv>,
-    pub ciph: Vec<u8>,
-    pub created_at: time::OffsetDateTime,
-}
 
 pub async fn load_decrypted_deks(
     deps: &impl Db,
     is_leader: IsLeaderDb,
     secrets: &dyn AuthlySecrets,
 ) -> anyhow::Result<DecryptedDeks> {
-    let mut opt_master_version = cryptography_db::load_cr_master_version(deps).await?;
+    let mut opt_master_version = crypto_repo::load_cr_master_version(deps).await?;
 
     let deks = if is_leader.0 {
         match opt_master_version {
             None => {
                 let decrypted = gen_new_master(secrets).await?;
-                cryptography_db::insert_cr_master_version(deps, decrypted.encrypted.clone())
-                    .await?;
+                crypto_repo::insert_cr_master_version(deps, decrypted.encrypted.clone()).await?;
 
                 gen_prop_deks(deps, &decrypted, is_leader).await?
             }
@@ -84,7 +41,7 @@ pub async fn load_decrypted_deks(
                     info!("waiting for leader to initialize master version");
                     tokio::time::sleep(Duration::from_secs(1)).await;
 
-                    opt_master_version = cryptography_db::load_cr_master_version(deps).await?;
+                    opt_master_version = crypto_repo::load_cr_master_version(deps).await?;
                 }
             }
         };
@@ -129,53 +86,6 @@ async fn decrypt_master(
         encrypted,
         key: AesKey::new((*secret.expose_secret()).into()),
     })
-}
-
-pub async fn gen_prop_deks(
-    deps: &impl Db,
-    decrypted_master: &DecryptedMaster,
-    is_leader: IsLeaderDb,
-) -> anyhow::Result<HashMap<PropId, AesKey>> {
-    let old_encrypted_deks = cryptography_db::list_all_cr_prop_deks(deps).await?;
-    let mut new_encrypted_deks: HashMap<PropId, EncryptedDek> = Default::default();
-    let mut decrypted_deks: HashMap<PropId, AesKey> = Default::default();
-
-    for id in all_encrypted_props() {
-        let decrypted_dek = if let Some(encrypted) = old_encrypted_deks.get(&id) {
-            let nonce = encrypted.nonce;
-            let decrypted_dek = decrypted_master
-                .key
-                .aes()
-                .decrypt(&nonce, encrypted.ciph.as_ref())?;
-
-            AesKey::load(Zeroizing::new(decrypted_dek))?
-        } else {
-            let nonce = random_nonce();
-            let mut dek = Aes256GcmSiv::generate_key(OsRng);
-            let ciph = decrypted_master.key.aes().encrypt(&nonce, dek.as_slice())?;
-
-            new_encrypted_deks.insert(
-                id.into(),
-                EncryptedDek {
-                    nonce,
-                    ciph,
-                    created_at: time::OffsetDateTime::now_utc(),
-                },
-            );
-
-            let key = AesKey::new(dek);
-            dek.zeroize();
-            key
-        };
-
-        decrypted_deks.insert(id.into(), decrypted_dek);
-    }
-
-    if is_leader.0 && !new_encrypted_deks.is_empty() {
-        cryptography_db::insert_cr_prop_deks(deps, new_encrypted_deks).await?;
-    }
-
-    Ok(decrypted_deks)
 }
 
 fn all_encrypted_props() -> impl Iterator<Item = BuiltinProp> {
