--- src/test_support/mod.rs
@@ -1,3 +1,0 @@
-mod test_ctx;
-
-pub use test_ctx::TestCtx;
