+++ src/tests/test_ultradb.rs
@@ -4,10 +4,11 @@ use authly_common::{
     id::ServiceId,
     proto::service::{self as proto, authly_service_client::AuthlyServiceClient},
 };
+use authly_test::test_ctx::TestCtx;
 use hexhex::hex_literal;
 
 use super::{compile_and_apply_doc_dir, tonic_request};
-use crate::{proto::service_server::AuthlyServiceServerImpl, test_support::TestCtx};
+use crate::proto::service_server::AuthlyServiceServerImpl;
 
 const ULTRADB: ServiceId =
     ServiceId::from_raw_array(hex_literal!("ec29ba1d23cb43f89b7c73db6f177a1d"));
